Use a thread to write compilation results
We spend roughly half of the time of codegen writing to files. This
process is fairly independent from the rest of the codegen stage. Thus,
we can easily parallelize it, and this change does exactly that.

We use a secondary thread to write the goto model and all the other
json files.