Run benchcomp if PR is labeled 'Z-BenchCI'
The "Auto Label" workflow correctly label our PRs; however, this action
do not trigger a PR 'labeled' workflow. To automatically do that, we
now trigger the benchmark workflow from the labeler workflow.

In this PR, we also change the workflow to run after any update to a PR
that has the 'Z-BenchCI' label.